Fix page css issues
Task #27
This change updates general page css to be containers, as
well as use the css variable for height.  Along with this
much of the changes here are for the business page and
its spacing.  These changes support better mobile UX
and put the footer to the bottom of the page